wip: components/App refactor
