+++ components/App/Loader.scss
