+++ lib/index.tsx
@@ -2,7 +2,7 @@ import './bugsnag.js'; // Initialize ASAP!
 
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App, { setActivityForApp } from '../components/App.js';
+import App, { setActivityForApp } from '../components/App/App.js';
 import LoadActivity from './LoadActivity.js';
 import { syncPackagesHash } from './ModuleCache.js';
 import { setActivityForRequestCache } from './fetchJSON.js';
